cli: add "headless" mode, fixes #34
This patch adds a --headless argument to the game server CLI, that
disables the GUI.

It's a little bit hacky—the CLI args are passed to the Application
constructor, by bypassing the Application singleton's `getInstance`
method.

An alternative solution would be to parse the CLI args into (a)
static `options` variable(s), and have the Application instance
use that. Then the options are global but we'd be agnostic about
the existence of the Application instance in the `main()` method.